Discord Bot/requirements.txt
discord.py==1.6.0
SQLAlchemy==1.4.3
psycopg2_binary==2.8.6
pandas==1.1.5
discord==1.0.1
psycopg2==2.8.6
python-dotenv==0.16.0
